+++ src/components/FileSection/index.js
@@ -1,0 +1,1 @@
+export { default } from './FileSection';
